requirements.txt
torch
transformers
typing
lumibot
alpaca-py
alpaca-trade-api
alpha_vantage
lib-pybroker
bs4